+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/option/SoraSpotlightOption.kt
@@ -8,7 +8,7 @@ class SoraSpotlightOption {
     /**
      * サイマルキャストの rid
      */
-    var simulcastRid: SoraSimulcastOption.SimulcastRid? = null
+    var simulcastRid: SoraVideoOption.SimulcastRid? = null
 
     /**
      * スポットライト機能のアクティブな配信数を指定します