+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/Catalog.kt
@@ -1,8 +1,6 @@
 package jp.shiguredo.sora.sdk.channel.signaling.message
 
-import android.os.Build
 import com.google.gson.annotations.SerializedName
-import jp.shiguredo.sora.sdk.BuildConfig
 import jp.shiguredo.sora.sdk.util.SDKInfo
 
 data class MessageCommonPart(
@@ -31,7 +29,7 @@ data class ConnectMessage(
         @SerializedName("spotlight")   var spotlight:               Any? = null,
         @SerializedName("spotlight_number") var spotlightNumber: Int? = null,
         @SerializedName("simulcast")   var simulcast:               Boolean? = false,
-        @SerializedName("simulcast_rid") var simulcastRid: SimulcastRid? = null,
+        @SerializedName("simulcast_rid") var simulcastRid: String? = null,
         @SerializedName("video")       var video:                   Any? = null,
         @SerializedName("audio")       var audio:                   Any? = null,
         @SerializedName("sora_client") val soraClient:              String = SDKInfo.sdkInfo(),
@@ -62,29 +60,6 @@ data class OpusParams(
         @SerializedName("useinbandfec")    var useinbandfec:    Boolean? = null,
         @SerializedName("usedtx")          var usedtx:          Boolean? = null
 )
-
-/**
- * サイマルキャストの rid
- */
-enum class SimulcastRid(val value: String) {
-
-    /**
-     * r0
-     */
-    R0("r0"),
-
-    /**
-     * r1
-     */
-    R1("r1"),
-
-    /**
-     * r2
-     */
-    R2("r2");
-
-    override fun toString(): String = value
-}
 
 data class IceServer(
         @SerializedName("urls")       val urls:       List<String>,