+++ CHANGES.md
@@ -9,6 +9,13 @@
 - FIX
     - バグ修正
 
+## 2021.1.1
+
+- [FIX] Sora への接続時に simulcast_rid を指定するとエラーになる現象を修正する
+  - @enm10k
+- [CHANGE] enum class SimulcastRid の定義を `jp.shiguredo.sora.sdk.channel.signaling.message` から `jp.shiguredo.sora.sdk.channel.option.SoraVideoOption` に移動する
+  - @enm10k
+
 ## 2021.1
 
 - [CHANGE] SoraAudioOption.Codec から PCMU を外す